[#1383] Implement utility as CLI command
