+++ bin/ots
@@ -72,6 +72,10 @@ class OT::CLI::Definition
   option :d, :domain, String, "Domain to revalidate"
   option :c, :custid, String, "Customer ID to revalidate"
   command revalidate_domains: OT::CLI
+
+  usage 'ots change-email OLD_EMAIL NEW_EMAIL [REALM] [DOMAIN OLD_ID]...'
+  about 'Change customer email address and update all related records'
+  command change_email: OT::CLI
 end
 
 begin
